examples/python/saga-loyalty-earn/pyproject.toml
[project]
name = "saga-loyalty-earn"
version = "0.1.0"
description = "Loyalty earn saga gRPC server for angzarr examples"
requires-python = ">=3.11"
dependencies = [
    "protobuf>=5.0.0",
    "grpcio>=1.60.0",
    "grpcio-health-checking>=1.60.0",
    "structlog>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["server.py"]
